refactor: 댓글 관련 컴포넌트 가독성 개선
